+++ CMakeLists.txt
@@ -1,7 +1,7 @@
 idf_component_register(
   SRCS "src/esp_dmx.c" "src/dmx/uart.c" "src/dmx/timer.c" "src/dmx/nvs.c"
-       "src/dmx/hal.c" "src/rdm/controller.c" "src/rdm/responder.c"
-       "src/rdm/parameters.c" "src/rdm/utils.c"
+       "src/dmx/hal.c" "src/dmx/utils.c" "src/rdm/controller.c"
+       "src/rdm/responder.c" "src/rdm/parameters.c"
   INCLUDE_DIRS "src" 
   REQUIRES driver esp_timer esp_common esp_hw_support nvs_flash
 )
