+++ src/dmx/struct.h
@@ -11,11 +11,11 @@
 
 #include "dmx/caps.h"
 #include "dmx/types.h"
+#include "dmx/utils.h"
 #include "freertos/queue.h"
 #include "freertos/semphr.h"
 #include "rdm/responder.h"
 #include "rdm/types.h"
-#include "rdm/utils.h"
 
 #ifdef __cplusplus
 extern "C" {
