+++ src/dmx/utils.c
